moved MapGenerationListeners to packages they are based upon
ProgressListener only needs a Property which can globally be used. So put it to the interface and map generation.
PreviewGenerationListener belongs to GameWorld and now resides in the same package.
